refactor shared directory
