+++ apps/server/src/core/error/filter/global-error.filter.ts
@@ -1,8 +1,8 @@
 import { ArgumentsHost, Catch, ExceptionFilter, HttpException, InternalServerErrorException } from '@nestjs/common';
 import * as _ from 'lodash';
 import { Response } from 'express';
-import { BusinessError } from '@shared/error/business.error';
-import { ApiValidationError } from '@shared/error/api-validation.error';
+import { BusinessError } from '@shared/common/error/business.error';
+import { ApiValidationError } from '@shared/common/error/api-validation.error';
 import { Logger } from '../../logger/logger.service';
 import { ErrorResponse } from '../dto/error.response';
 import { FeathersError } from '../interface';
