+++ apps/server/src/modules/task/uc/task.uc.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { EntityId, IPagination } from '@shared/domain';
-import { Counted } from '@shared/types';
+import { Counted } from '@shared/domain/types';
 import { TaskRepo } from '../repo/task.repo';
 import { Task } from '../entity';
