+++ apps/server/src/server.controller.ts
@@ -1,13 +1,10 @@
 import { Controller, Get } from '@nestjs/common';
-import { ServerService } from './server.service';
 
 @Controller()
 export class ServerController {
-	constructor(private readonly serverService: ServerService) {}
-
 	/** default route to test public access */
 	@Get()
 	getHello(): string {
-		return this.serverService.getHello();
+		return 'Schulcloud Server API';
 	}
 }
