+++ apps/server/src/server.module.ts
@@ -4,7 +4,6 @@ import { Dictionary, IPrimaryKey } from '@mikro-orm/core';
 import { AuthModule } from './modules/authentication/auth.module';
 import { NewsModule } from './modules/news/news.module';
 import { ServerController } from './server.controller';
-import { ServerService } from './server.service';
 import { DB_URL } from './config';
 import { TaskModule } from './modules/task/task.module';
 import { CourseNews, News, SchoolInfo, SchoolNews, TeamNews, UserInfo } from './modules/news/entity';
@@ -44,6 +43,5 @@ import { TeamInfo } from './modules/news/entity/team-info.entity';
 		CoreModule,
 	],
 	controllers: [ServerController],
-	providers: [ServerService],
 })
 export class ServerModule {}
