--- apps/server/src/server.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class ServerService {
-	getHello(): string {
-		return 'Hello World!';
-	}
-}
