+++ apps/server/src/shared/common/error/index.ts
@@ -1,0 +1,2 @@
+export * from './api-validation.error';
+export * from './business.error';
