+++ apps/server/src/shared/common/index.ts
@@ -1,0 +1,2 @@
+export * from './error';
+export * from './interceptor';
