+++ apps/server/src/shared/controller/dto/index.ts
@@ -1,1 +1,2 @@
 export * from './pagination.query';
+export * from './pagination.response';
