+++ apps/server/src/shared/controller/index.ts
@@ -1,1 +1,3 @@
 export * from './dto';
+export * from './pipe';
+export * from './transformer';
