+++ apps/server/src/shared/domain/index.ts
@@ -1,3 +1,3 @@
 export * from './entity';
-export * from './entity-id';
 export * from './interface';
+export * from './types';
