+++ apps/server/src/shared/domain/types/counted.ts
