+++ apps/server/src/shared/domain/types/entity-id.ts
@@ -1,1 +1,4 @@
+/**
+ * Entity id type
+ */
 export type EntityId = string;
