+++ apps/server/src/shared/domain/types/sorting.ts
@@ -1,0 +1,10 @@
+import { ValueOf } from './value-of';
+
+export const SortOrder = {
+	asc: 'asc',
+	desc: 'desc',
+} as const;
+
+export type SortOrderValue = ValueOf<typeof SortOrder>;
+
+export type SortOrderMap<T> = Partial<Record<keyof T, SortOrderValue>>;
