+++ apps/server/src/shared/domain/types/value-of.ts
@@ -1,0 +1,4 @@
+/**
+ * The type of the values in T (counterpart to keyof)
+ */
+export type ValueOf<T> = T[keyof T];
