+++ apps/server/test/app.e2e-spec.ts
@@ -20,6 +20,6 @@ describe('ServerController (e2e)', () => {
 	});
 
 	it('/ (GET)', () => {
-		return request(app.getHttpServer()).get('/').expect(200).expect('Hello World!');
+		return request(app.getHttpServer()).get('/').expect(200).expect('Schulcloud Server API');
 	});
 });
